🚚 버튼, 텍스트 폴더로 이동 #200
컴포너트 폴더화
