+++ client/src/components/atoms/Buttons/TransparentButton/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styled from '@emotion/styled';
+import { StyledTransparentButton } from './style';
 
 interface IProps {
   onClick: (event: React.MouseEvent<HTMLButtonElement>) => void;
@@ -7,20 +7,6 @@ interface IProps {
   zIdx?: string;
 }
 
-interface IStyledProps {
-  margin: string;
-  zIdx: string;
-}
-
-const StyledTransparentButton = styled.button<IStyledProps>`
-  border: none;
-  background: transparent;
-  margin: ${({ margin }) => margin};
-  z-index: ${({ zIdx }) => zIdx};
-  padding: 0;
-  width: fit-content;
-  cursor: pointer;
-`;
 const TransparentButton: React.FC<IProps> = ({ children, onClick, margin = '0', zIdx = '0' }) => {
   return (
     <StyledTransparentButton onClick={onClick} margin={margin} zIdx={zIdx}>
